requirements.txt
# general
git+git://github.com/Mirantis/stepler.git
paramiko==2.0.2
pytest==3.0.2
pytest-xdist==1.15.0
requests==2.11.1
six==1.10.0
waiting==1.3.0
subprocess32==3.2.7; python_version < '3.2' and os_name == 'posix'
mock==2.0.0
pyaml==16.9.0
python-dateutil==2.5.3
simplejson==3.10.0

# ui testing
attrdict==2.0.0
PyHamcrest==1.9.0
python-pom==1.0.2
selenium==3.0.1
xvfbwrapper==0.2.8

# debug
ipdb==0.10.1
ipython==5.1.0

# doc
Sphinx==1.4.6
oslosphinx==4.7.0

# verification
tox==2.4.1
